// generated/models/MemeClaim.ts
export class MemeClaim {
    'drop_id': string;
    'meme_id': number;
    'season': number;
    'image_location'?: string | null;
    'animation_location'?: string | null;
    'metadata_location'?: string | null;
    'arweave_synced_at'?: number | null;
    'media_uploading'?: boolean;
    'edition_size'?: number | null;
    'description': string;
    'name': string;
    'image_url'?: string | null;
    'attributes': Array<MemeClaimAttribute>;
    'image_details'?: MemeClaimImageDetails;
    'animation_url'?: string | null;
    'animation_details'?: MemeClaimAnimationDetails | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "drop_id",
            "baseName": "drop_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "meme_id",
            "baseName": "meme_id",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "season",
            "baseName": "season",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "image_location",
            "baseName": "image_location",
            "type": "string",
            "format": ""
        },
        {
            "name": "animation_location",
            "baseName": "animation_location",
            "type": "string",
            "format": ""
        },
        {
            "name": "metadata_location",
            "baseName": "metadata_location",
            "type": "string",
            "format": ""
        },
        {
            "name": "arweave_synced_at",
            "baseName": "arweave_synced_at",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "media_uploading",
            "baseName": "media_uploading",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "edition_size",
            "baseName": "edition_size",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "image_url",
            "baseName": "image_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "Array<MemeClaimAttribute>",
            "format": ""
        },
        {
            "name": "image_details",
            "baseName": "image_details",
            "type": "MemeClaimImageDetails",
            "format": ""
        },
        {
            "name": "animation_url",
            "baseName": "animation_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "animation_details",
            "baseName": "animation_details",
            "type": "MemeClaimAnimationDetails",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return MemeClaim.attributeTypeMap;
    }

    public constructor() {
    }
}